ENH improve dynamic module loading + clean up
